requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-cors-headers==3.0.1
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==2.8
kombu==4.5.0
oauthlib==3.0.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
Serializer==0.2.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
vine==1.3.0
